+++ Example/DemoTests/TestView.h
@@ -1,0 +1,12 @@
+//  
+//  Copyright (c) 2014 Aleksander Zubala All rights reserved.
+//  
+
+
+#import <Foundation/Foundation.h>
+
+
+@interface TestView : UIView
+@property(nonatomic, strong) UIView *subview;
+@property(nonatomic, strong) NSNumber *metric;
+@end